+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Borderless Topic view for more characters of topic names
+- Improved mouse scroll
 - Display broker text easier to read
 - Display version & broker in the lower right corner
 - Display MQTT connection error in its own area